% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{HammingDistance}
\alias{HammingDistance}
\title{Hamming (p-)distance between discrete character rows}
\usage{
HammingDistance(X)
}
\arguments{
\item{X}{A taxa x characters matrix of state labels.}
}
\value{
A symmetric taxa x taxa distance matrix.
}
\description{
Proportion of characters differing between each pair of taxa; valid for
any state-space size (unlike a binary Hamming distance, no assumption is
made about the number of states a character can take).
}
