% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimony.R
\name{FindTNT}
\alias{FindTNT}
\title{Locate the TNT executable}
\usage{
FindTNT()
}
\value{
Path to the TNT executable.
}
\description{
Looks for a TNT installation, in order: the \code{phyloPCA.tnt} option, the
\code{TNT_PATH} environment variable, then a handful of common install
locations.
}
