% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloPCA-package.R
\docType{package}
\name{phyloPCA-package}
\alias{phyloPCA}
\alias{phyloPCA-package}
\title{phyloPCA: Phylogenetic Signal in Continuous, Discrete and Geometric-Morphometric Data}
\description{
Simulates continuous, discrete and geometric-morphometric ('shape') character data on a fixed generative tree, and reconstructs phylogeny from each using neighbour-joining, maximum parsimony and Bayesian inference, so that reconstruction accuracy can be compared across data types and methods. Provides the simulation, alignment, tree-inference and scoring functions, together with the cached results, underlying Smith (2026), a reply to Raskin et al. (2026) \doi{10.1002/ajpa.70306} on whether principal components analysis can recover phylogenetic signal from continuous morphological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/phylo-pca/phylo-pca.github.io}
  \item \url{https://phylo-pca.github.io}
  \item Report bugs at \url{https://github.com/phylo-pca/phylo-pca.github.io/issues}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID})

Authors:
\itemize{
  \item Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID})
}

}
\keyword{internal}
